+++ packages/react/src/combobox/combobox-option-group.tsx
@@ -3,28 +3,21 @@ import { mergeProps } from '@zag-js/react'
 import { forwardRef, type ComponentPropsWithoutRef } from 'react'
 import { createSplitProps } from '../create-split-props'
 import { ark } from '../factory'
-import { type Assign } from '../types'
+import type { Assign } from '../types'
 import { useComboboxContext } from './combobox-context'
 
 export type ComboboxOptionGroupProps = Assign<
-  ComponentPropsWithoutRef<typeof ark.ul>,
+  ComponentPropsWithoutRef<typeof ark.div>,
   OptionGroupProps
 >
 
-export const ComboboxOptionGroup = forwardRef<HTMLUListElement, ComboboxOptionGroupProps>(
+export const ComboboxOptionGroup = forwardRef<HTMLDivElement, ComboboxOptionGroupProps>(
   (props, ref) => {
-    const [optionProps, { children, ...liProps }] = createSplitProps<OptionGroupProps>()(props, [
-      'label',
-    ])
-
+    const [optionProps, localProps] = createSplitProps<OptionGroupProps>()(props, ['id'])
     const { getOptionGroupProps } = useComboboxContext()
-    const mergedProps = mergeProps(getOptionGroupProps(optionProps), liProps)
+    const mergedProps = mergeProps(getOptionGroupProps(optionProps), localProps)
 
-    return (
-      <ark.ul {...mergedProps} ref={ref}>
-        {children ? children : optionProps.label}
-      </ark.ul>
-    )
+    return <ark.div {...mergedProps} ref={ref} />
   },
 )