+++ packages/react/src/combobox/combobox.stories.tsx
@@ -10,6 +10,7 @@ import {
   ComboboxLabel,
   ComboboxOption,
   ComboboxOptionGroup,
+  ComboboxOptionGroupLabel,
   ComboboxPositioner,
   ComboboxTrigger,
   type ComboboxOptionProps,
@@ -57,7 +58,10 @@ export const Basic = () => {
           <Portal>
             <ComboboxPositioner>
               <ComboboxContent>
-                <ComboboxOptionGroup label="Javascript">
+                <ComboboxOptionGroup id="framework">
+                  <ComboboxOptionGroupLabel htmlFor="framework">
+                    Frameworks
+                  </ComboboxOptionGroupLabel>
                   {options.map((item, index) => (
                     <ComboboxOption
                       key={`${item.value}:${index}`}