+++ packages/react/src/select/select-option-group.tsx
@@ -1,19 +1,22 @@
 import { mergeProps } from '@zag-js/react'
+import type { OptionGroupProps } from '@zag-js/select'
 import { forwardRef, type ComponentPropsWithoutRef } from 'react'
+import { createSplitProps } from '../create-split-props'
 import { ark } from '../factory'
 import type { Assign } from '../types'
 import { useSelectContext } from './select-context'
 
 export type SelectOptionGroupProps = Assign<
   ComponentPropsWithoutRef<typeof ark.div>,
-  { id: string }
+  OptionGroupProps
 >
 
 export const SelectOptionGroup = forwardRef<HTMLDivElement, SelectOptionGroupProps>(
   (props, ref) => {
-    const { id, ...divProps } = props
+    const [optionProps, localProps] = createSplitProps<OptionGroupProps>()(props, ['id'])
+
     const { getOptionGroupProps } = useSelectContext()
-    const mergedProps = mergeProps(getOptionGroupProps({ id }), divProps)
+    const mergedProps = mergeProps(getOptionGroupProps(optionProps), localProps)
 
     return <ark.div {...mergedProps} ref={ref} />
   },