+++ packages/vue/src/combobox/combobox-option-group-label.tsx
@@ -1,0 +1,28 @@
+import { defineComponent, type PropType } from 'vue'
+import { ark, type HTMLArkProps } from '../factory'
+import { getValidChildren, type ComponentWithProps } from '../utils'
+import { useComboboxContext } from './combobox-context'
+
+export interface ComboboxOptionGroupLabelProps extends HTMLArkProps<'label'> {
+  htmlFor: string
+}
+
+export const ComboboxOptionGroupLabel: ComponentWithProps<ComboboxOptionGroupLabelProps> =
+  defineComponent({
+    name: 'ComboboxOptionGroupLabel',
+    props: {
+      htmlFor: String as PropType<ComboboxOptionGroupLabelProps['htmlFor']>,
+    },
+    setup(props, { slots, attrs }) {
+      const api = useComboboxContext()
+
+      return () => (
+        <ark.label
+          {...api.value.getOptionGroupLabelProps({ htmlFor: props.htmlFor as string })}
+          {...attrs}
+        >
+          {() => getValidChildren(slots)}
+        </ark.label>
+      )
+    },
+  })